phone number to bigint

+++ create.sql
@@ -2,7 +2,7 @@ CREATE TABLE rideshare_user
 (netid VARCHAR(7) NOT NULL UNIQUE PRIMARY KEY,
  name VARCHAR(256) NOT NULL,
  duke_email VARCHAR(256) NOT NULL,
- phone_number INTEGER NOT NULL,
+ phone_number BIGINT NOT NULL,
  affiliation VARCHAR(256) NOT NULL CHECK(affiliation IN ('Undergraduate', 'Graduate')),
  school VARCHAR(256) NOT NULL CHECK(school IN ('Trinity', 'Pratt', 'Nicolas', 'Law', 'Fuqua', 
                 'Nursing', 'Medicine', 'Other')),
